Updated sidebar structure
