+++ dictionary/src/common/components/navigation/definitions/Sidebar.tsx
@@ -4,11 +4,11 @@ interface SidebarProps {
 }
 
 const Sidebar = ({ word, sources }: SidebarProps) => {
-    return <div className=" h-full w-52 fixed left-0 top-[31px] bg-gray-100 pl-8 pt-8">
+    return <nav className=" h-full w-52 fixed left-0 top-[31px] bg-gray-100 pl-8 pt-8">
         
         {/* Title */}
-        <h1 className="text-xl text-black font-bold">{ word }</h1>
-    </div>
+        <h2 className="text-xl text-black font-bold">{ word }</h2>
+    </nav>
 }
 
 export default Sidebar;
